refactor(eslint): move shared call-shape helpers to a leaf module
sweep-binding.mjs needed calledName/SQL_SINKS from require-table-teardown.mjs,
which imports the resolver back. Move calledName, staticString and SQL_SINKS to
eslint/sql-call-shapes.mjs so nothing imports in a cycle, and pin that a fixed
name held in a variable is not sweep-bound.
